% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{eff.cjdesign}
\alias{eff.cjdesign}
\title{Print design efficiency}
\usage{
\method{eff}{cjdesign}(object, ...)
}
\arguments{
\item{object}{is an object of class \code{cjdesign}, a data frame returned
from the \code{makeDoe()} function.}
}
\value{
For a given design, prints the D efficiency and whether the design
is balanced and / or orthogonal.
}
\description{
For a given design, prints the D efficiency and whether the design is
balanced and / or orthogonal.
}
\examples{
# Define the attributes and levels of an experiment
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c('Fuji', 'Gala', 'Honeycrisp', 'Pink Lady', 'Red Delicious'),
  freshness = c('Excellent', 'Average', 'Poor')
)

# Make a fractional-factorial design of experiment
doe <- makeDoe(levels, type = "D", nTrials = 15)

# View the summary information about the design
eff(doe)
}
\keyword{logitr}
\keyword{wtp}
