% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{eff}
\alias{eff}
\title{Print design D and A efficiency}
\usage{
eff(object)
}
\arguments{
\item{object}{is an object of class \code{cjdesign}, a data frame returned
from the \code{makeDoe()} function.}
}
\value{
Prints the D and A efficiencies of a design.
}
\description{
Prints the D and A efficiencies of a design.
}
\keyword{logitr}
\keyword{wtp}
