% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateModels.R
\name{estimateModels}
\alias{estimateModels}
\title{Estimate the same model on different size subsets of a data set}
\usage{
estimateModels(
  nbreaks = 10,
  data,
  parNames,
  choiceName = "choice",
  obsIDName = "obsID",
  priceName = NULL,
  randPars = NULL,
  randPrice = NULL,
  modelSpace = "pref",
  weightsName = NULL,
  clusterName = NULL,
  robust = FALSE,
  options = list()
)
}
\arguments{
\item{nbreaks}{The number of different sample size groups.}

\item{data}{A data frame containing choice observations.}

\item{parNames}{The names of the parameters to be estimated in the model.
Must be the same as the column names in the \code{data} argument. For WTP space
models, do not include price in \code{parNames}.}

\item{choiceName}{The name of the column that identifies the choice variable.}

\item{obsIDName}{The name of the column that identifies each choice
observation.}

\item{priceName}{The name of the column that identifies the price variable.
Only required for WTP space models. Defaults to \code{NULL}.}

\item{randPars}{A named vector whose names are the random parameters and
values the distribution: \code{'n'} for normal or \code{'ln'} for log-normal.
Defaults to \code{NULL}.}

\item{randPrice}{The random distribution for the price parameter: \code{'n'} for
normal or \code{'ln'} for log-normal. Only used for WTP space MXL models.
Defaults to \code{NULL}.}

\item{modelSpace}{Set to \code{'wtp'} for WTP space models. Defaults to \code{"pref"}.}

\item{weightsName}{The name of the column that identifies the weights to be
used in model estimation. Optional. Defaults to \code{NULL}.}

\item{clusterName}{The name of the column that identifies the cluster
groups to be used in model estimation. Optional. Defaults to \code{NULL}.}

\item{robust}{Determines whether or not a robust covariance matrix is
estimated. Defaults to \code{FALSE}. Specification of a clusterName or
weightsName will override the user setting and set this to `TRUE' (a
warning will be displayed in this case). Replicates the functionality of
Stata's cmcmmixlogit.}

\item{options}{A list of options to control the model estimation. See the
\code{options} argument in the {logitr} package for details.}
}
\value{
Returns a nested data frame with each estimated model object in
the \code{model} column.
}
\description{
This function estimates the same model multiple times using different size
subsets of a set of choice data. The number of models to run is set by the
\code{nbreaks} argument, which breaks up the data into groups of increasing
sample sizes. All models are estimated models using the {logitr} package.
}
\examples{
\dontrun{
library(conjointTools)

# Define the attributes and levels
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c('Fuji', 'Gala', 'Honeycrisp', 'Pink Lady', 'Red Delicious'),
  freshness = c('Excellent', 'Average', 'Poor')
)

# Make a full-factorial design of experiment and recode the levels
doe <- makeDoe(levels)
doe <- recodeDesign(doe, levels)

# Make the survey
survey <- makeSurvey(
    doe       = doe,  # Design of experiment
    nResp     = 2000, # Total number of respondents (upper bound)
    nAltsPerQ = 3,    # Number of alternatives per question
    nQPerResp = 6     # Number of questions per respondent
)

# Simulate random choices for the survey
data <- simulateChoices(
    survey    = survey,
    altIDName = "altID",
    obsIDName = "obsID"
)

# Estimate models with different sample sizes
models <- estimateModels(
    nbreaks     = 10,
    data        = data,
    parNames    = c("price", "type", "freshness"),
    choiceName  = "choice",
    obsIDName   = "obsID"
)
}
}
\keyword{logit,}
\keyword{logitr,}
\keyword{mnl,}
\keyword{mxl,}
\keyword{power}
\keyword{sample}
\keyword{size,}
