% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPower.R
\name{getPower}
\alias{getPower}
\title{Estimate errors from conjoint designs of different sample sizes}
\usage{
getPower(
  survey,
  parNames = NULL,
  parTypes = NULL,
  interactions = FALSE,
  nbreaks = 10,
  randPars = NULL
)
}
\arguments{
\item{survey}{The choice survey data frame exported from the \code{makeSurvey()} function.}

\item{parNames}{A vector of the names of the parameters to be estimated in the model. Must be the same as the column names in the \code{survey} argument.}

\item{parTypes}{A vector determining the type of each variable: "c" for continuous, or "d" for discrete. Continuous variables will be linearly coded whereas discrete variables will be dummy coded with one level removed for identification. Defaults to \code{NULL}, in which case all parameters are coded as "d" for discrete.}

\item{interactions}{A logical value for adding interactions between all parameters in \code{parNames}. Defaults to \code{FALSE}.}

\item{nbreaks}{The number of different sample size groups.}

\item{randPars}{A named vector whose names are the random parameters and values the distribution: \code{'n'} for normal or \code{'ln'} for log-normal. Defaults to \code{NULL}.}
}
\value{
Returns a data frame of the standard error values for different sample sizes.
}
\description{
This function estimates multiple models with increasing sample sizes of a
particular survey design. The number of models to run is set by the
\code{nbreaks} argument, which breaks up the survey design into groups of
increasing sample size. This function returns estimated models from the
{logitr} package. View a summary of the results with the \code{summary()}
function.
}
\examples{
\dontrun{
library(conjointTools)
}
}
\keyword{logit,}
\keyword{logitr,}
\keyword{mnl,}
\keyword{mxl,}
\keyword{sample}
\keyword{size}
