% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{makeSurvey}
\alias{makeSurvey}
\title{Make a conjoint survey from a design of experiment}
\usage{
makeSurvey(doe, nResp, nAltsPerQ, nQPerResp)
}
\arguments{
\item{doe}{The design of experiment returned from the \code{makeDoe()} function}

\item{nResp}{Maximum number of survey respondents}

\item{nAltsPerQ}{Number of alternatives per question}

\item{nQPerResp}{Number of questions per respondent}
}
\value{
Returns a coded conjoint survey by sampling from a design of
experiment returned from the \code{makeDoe()} function. The design eliminates the
possibility of duplicate alternatives appearing in the same choice question.
}
\description{
Make a conjoint survey from a design of experiment
}
\examples{
\dontrun{
# Generate a full factorial design of experiment about apples
doe <- makeDoe(
    levels = c(3, 3, 3),
    varNames = c("price", "type", "freshness"),
    type = "full"
)

# Make the coded conjoint survey
survey <- makeSurvey(
    doe       = doe,  # Design of experiment
    nResp     = 1000, # Total number of respondents (upper bound)
    nAltsPerQ = 3,    # Number of alternatives per question
    nQPerResp = 6     # Number of questions per respondent
)
}
}
