% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{miscmethods.cjdesign}
\alias{miscmethods.cjdesign}
\alias{eff.cjdesign}
\alias{print.cjdesign}
\title{Methods for cjdesign objects}
\usage{
\method{eff}{cjdesign}(object, ...)
}
\arguments{
\item{...}{further arguments.}

\item{x}{is an object of class \code{cjdesign}.}

\item{digits}{the number of digits for printing, defaults to \code{3}.}

\item{width}{the width of the printing.}
}
\description{
Miscellaneous methods for \code{cjdesign} class objects.
}
