% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateChoices.R
\name{randLN}
\alias{randLN}
\title{Define "true" model parameters as normally-distributed.}
\usage{
randLN(mu = 0, sigma = 1)
}
\arguments{
\item{mu}{Mean of the distribution on the log scale, defaults to \code{0}.}

\item{sigma}{Standard deviation of the distribution on the log scale,
defaults to \code{1}.}
}
\value{
A list defining log-normally-distributed parameters of the "true"
utility model used to simulate choices in the \code{simulateChoices()} function.
}
\description{
Define "true" model parameters as normally-distributed. Used in the
\code{simulateChoices()} function.
}
\examples{
\dontrun{
library(conjointTools)
}
}
