% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateChoices.R
\name{randN}
\alias{randN}
\title{Define "true" model parameters as normally-distributed.}
\usage{
randN(mu = 0, sigma = 1)
}
\arguments{
\item{mu}{Vector of means, defaults to \code{0}.}

\item{sigma}{Vector of standard deviations, defaults to \code{1}.}
}
\value{
A list defining normally-distributed parameters of the "true"
utility model used to simulate choices in the \code{simulateChoices()} function.
}
\description{
Define "true" model parameters as normally-distributed. Used in the
\code{simulateChoices()} function.
}
\examples{
\dontrun{
library(conjointTools)
}
}
