% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizer.R
\name{sampleSizePlot}
\alias{sampleSizePlot}
\title{Creates a plot of the sample size results.}
\usage{
sampleSizePlot(df)
}
\arguments{
\item{df}{The data frame of the output from the sampleSizer function.}
}
\value{
Creates a plot of the standard errors for different sample sizes.
}
\description{
Creates a plot of the sample size results.
}
\examples{
\dontrun{
# First run the sampleSizer function
sizeTest <- sampleSizer(
    data       = yogurt,
    obsIDName  = 'obsID',
    parNames   = c('price', 'feat', 'dannon', 'hiland', 'yoplait'),
    nbreaks    = 10,
    plot       = TRUE)

# Plot the results
sampleSizePlot(sizeTest)
}
}
