% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe.R
\name{makeDoe}
\alias{makeDoe}
\title{Make a design of experiment}
\usage{
makeDoe(ff, nResp, nAltsPerQ, nQPerResp)
}
\arguments{
\item{ff}{The full factorial design of experiment}

\item{nResp}{Total number of respondents (upper bound)}

\item{nAltsPerQ}{Number of alternatives per question}

\item{nQPerResp}{Number of questions per respondent}
}
\value{
Creates a design of experiment
}
\description{
Make a design of experiment
}
\examples{
\dontrun{
# Generate a full factorial design about apples
ff <- expand.grid(
    price     = seq(1, 4, 0.5), # $ per pound
    type      = c('Fuji', 'Gala', 'Honeycrisp', 'Pink Lady', 'Red Delicious'),
    freshness = c('Excellent', 'Average', 'Poor')
)

# Make the doe
doe <- makeDoe(
    ff        = ff,   # Full factorial design
    nResp     = 3000, # Total number of respondents (upper bound)
    nAltsPerQ = 3,    # Number of alternatives per question
    nQPerResp = 6     # Number of questions per respondent
)
}
}
