% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{makeSurvey}
\alias{makeSurvey}
\title{Make a conjoint survey from a design of experiment}
\usage{
makeSurvey(doe, nResp, nAltsPerQ, nQPerResp, group = NULL)
}
\arguments{
\item{doe}{A design of experiment data frame. Each row is an alternative,
each column is an attribute.}

\item{nResp}{Maximum number of survey respondents}

\item{nAltsPerQ}{Number of alternatives per question}

\item{nQPerResp}{Number of questions per respondent}

\item{group}{The name of the group column if your doe is designed with
groups (e.g. a labeled experiment where each set of alternatives has the
same type). Defaults to \code{NULL}.}
}
\value{
Returns a conjoint survey by randomly sampling from a design of
experiment data frame. The sampling is done iteratively to remove the
possibility of duplicate alternatives appearing in the same choice question.
}
\description{
Make a conjoint survey from a design of experiment
}
\examples{
\dontrun{
# Generate a full factorial design of experiment about apples
doe <- makeDoe(
    levels = c(3, 3, 3),
    varNames = c("price", "type", "freshness"),
    type = "full"
)

# Make the coded conjoint survey
survey <- makeSurvey(
    doe       = doe,  # Design of experiment
    nResp     = 1000, # Total number of respondents (upper bound)
    nAltsPerQ = 3,    # Number of alternatives per question
    nQPerResp = 6     # Number of questions per respondent
)
}
}
