% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizer.R
\name{sampleSizePlot}
\alias{sampleSizePlot}
\title{Creates a plot of the sample size results.}
\usage{
sampleSizePlot(df)
}
\arguments{
\item{df}{The data frame of the output from the \code{sampleSizer()} function.}
}
\value{
Returns a plot of the standard errors for different sample sizes.
}
\description{
Creates a plot of the sample size results.
}
\examples{
\dontrun{
}
}
