% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizer.R
\name{sampleSizer}
\alias{sampleSizer}
\title{Estimate standard errors from conjoint design}
\usage{
sampleSizer(
  survey,
  parNames = NULL,
  parTypes = NULL,
  nbreaks = 10,
  randPars = NULL,
  options = list(message = FALSE)
)
}
\arguments{
\item{survey}{The choice survey data frame exported from the \code{makeSurvey()} function.}

\item{parNames}{A vector of the names of the parameters to be estimated in the model. Must be the same as the column names in the \code{survey} argument.}

\item{parTypes}{A vector determining the type of each variable: "c" for continuous, or "d" for discrete. Continuous variables will be linearly coded whereas discrete variables will be dummy coded with one level removed for identification. Defaults to \code{NULL}, in which case all parameters are coded as discrete.}

\item{nbreaks}{The number of different sample size groups.}

\item{randPars}{A named vector whose names are the random parameters and values the distribution: \code{'n'} for normal or \code{'ln'} for log-normal. Defaults to \code{NULL}.}

\item{options}{A list of options to control the model estimation.}
}
\value{
Returns a data frame of the standard error values for different sample sizes.
}
\description{
This function allows you to estimate a model on a design of experiment
that you have not yet used to collect data. This allows you to learn
about deficiencies in your design of experiment and also assess the sample
size needed to achieve parameter precision levels before you go out and
use the design to collect data. The function fills out the survey with
random choices and estimates a model. It does this multiple times with an
increasing number of observations, set by the \code{nbreaks} argument. While the
coefficients in those models are meaningless, the \emph{standard errors} on the
coefficients are informative. The example below estimates 10 separate models
and then plots the standard errors against the number of observations. In
this example, assume that the yogurt data was not a completed survey but
rather a blank design of experiment with no observed choices.
}
\examples{
\dontrun{
}
}
\keyword{logit,}
\keyword{logitr,}
\keyword{mnl,}
\keyword{mxl,}
\keyword{sample}
\keyword{size}
